fix: export NgView and NgElement from element-registry
reverts breaking change
