+++ nativescript-angular/animations/animation-driver.ts
@@ -1,6 +1,6 @@
 import { AnimationPlayer } from "@angular/animations";
 
-import { NgView } from "../element-types";
+import { NgView } from "../element-registry";
 import { NativeScriptAnimationPlayer } from "./animation-player";
 import { Keyframe } from "./utils";
