+++ tests/app/tests/property-sets.ts
@@ -2,7 +2,11 @@
 import {assert} from "./test-config";
 import {View} from "ui/core/view";
 import {ViewUtil} from "nativescript-angular/view-util";
-import {NgView, ViewExtensions, ViewClassMeta} from "nativescript-angular/element-types";
+import {
+    NgView,
+    ViewExtensions,
+    ViewClassMeta,
+} from "nativescript-angular/element-registry";
 import {Red} from "color/known-colors";
 import {device, platformNames} from "platform";
 import {createDevice} from "./test-utils";
